requirements.txt
Django>=3.1.1
djangorestframework>=3.12.0
psycopg2-binary>=2.9.0
python-dotenv>=0.17.0 